% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_hysteresis_CF.R
\name{add_water_to_small}
\alias{add_water_to_small}
\title{Adds water to small depressions}
\usage{
add_water_to_small(
  current_storage = 0,
  delta_storage = 0,
  max_storage = 0,
  current_contrib_frac = 0,
  threshold = -0.01,
  iterations = 10
)
}
\arguments{
\item{current_storage}{}

\item{delta_storage}{}

\item{max_storage}{}

\item{current_contrib_frac}{}

\item{threshold}{}

\item{iterations}{}
}
\value{
Returns the updated depressional storage, contributing fraction and outflow
}
\description{
Adds water to small depressions
}
\examples{
\dontrun{
}

}
