% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{area_frac}
\alias{area_frac}
\title{Determines if area-volume estimation is by equation or rating}
\usage{
area_frac(volume, max_volume, max_area = NULL, rating_parameters)
}
\arguments{
\item{volume}{Volume of water in depressional storage}

\item{max_volume}{Volume of depression}

\item{max_area}{Area of depression}

\item{rating_parameters}{Parameters for estimating water area fraction}
}
\value{
Returns water area fraction
}
\description{
Determines if area-volume estimation is by equation or rating
}
\examples{
water_area_frac <- area_frac(1000, 2000, NULL, 1.2)
}
