% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{volfrac2areafrac_Clark}
\alias{volfrac2areafrac_Clark}
\title{Estimates total depressional water area fraction}
\usage{
volfrac2areafrac_Clark(volume_frac, p)
}
\arguments{
\item{volume_frac}{Fraction of depressional storage}

\item{p}{Exponent}
}
\value{
Returns the fractional water area
}
\description{
Estimates total depressional water area fraction
}
\examples{
areafrac <- volfrac2areafrac_Clark(0.5, 1.2)
}
